--- a/spec/controllers/answers_controller_with_conditional_questions_spec.rb
+++ b/spec/controllers/answers_controller_with_conditional_questions_spec.rb
@@ -11,41 +11,18 @@ RSpec.describe AnswersController, type: :controller do
     @section1, @section2, @section3 = template.sections
 
     # Different types of questions (than can have conditional options)
-    @conditional_questions = {
-      checkbox: create(:question, :checkbox, section: @section1, options: 5),
-      radiobutton: create(:question, :radiobuttons, section: @section2, options: 5),
-      dropdown: create(:question, :dropdown, section: @section3, options: 5)
-    }
+    @conditional_questions = create_conditional_questions
 
     # Questions that do not have conditional options for adding or removing
-    @non_conditional_questions = {
-      textarea: create_list(:question, 7, :textarea, section: @section1),
-      textfield: create_list(:question, 7, :textfield, section: @section2),
-      date: create_list(:question, 7, :date, section: @section3),
-      rda_metadata: create_list(:question, 7, :rda_metadata, section: @section1, options: 3),
-      checkbox: create_list(:question, 7, :checkbox, section: @section2, options: 3),
-      radiobutton: create_list(:question, 7, :radiobuttons, section: @section3, options: 3),
-      dropdown: create_list(:question, 7, :dropdown, section: @section1, options: 3),
-      multiselectbox: create_list(:question, 7, :multiselectbox, section: @section2, options: 3)
-    }
+    @non_conditional_questions = create_non_conditional_questions
 
     @plan = create(:plan, :creator, template: template)
     @user = @plan.owner
 
     # Answer the questions in List2
-    @answers = {}
-    @non_conditional_questions.each do |question_type, questions|
-      @answers[question_type] = questions.map do |question|
-        if %i[textarea textfield date rda_metadata].include?(question_type)
-          create(:answer, plan: @plan, question: question, user: @user)
-        else
-          create(:answer, plan: @plan, question: question, question_options: [question.question_options[2]], user: @user)
-        end
-      end
-    end
+    @answers = create_answers
 
     @all_questions_ids = (@conditional_questions.values + @non_conditional_questions.values.flatten).map(&:id)
-
     @all_answers_ids = @answers.values.flatten.map(&:id)
 
     sign_in(@user)
@@ -508,5 +485,43 @@ RSpec.describe AnswersController, type: :controller do
         end
       end
     end
+  end
+
+  private
+
+  def create_conditional_questions
+    {
+      checkbox: create(:question, :checkbox, section: @section1, options: 5),
+      radiobutton: create(:question, :radiobuttons, section: @section2, options: 5),
+      dropdown: create(:question, :dropdown, section: @section3, options: 5)
+    }
+  end
+
+  def create_non_conditional_questions
+    {
+      textarea: create_list(:question, 7, :textarea, section: @section1),
+      textfield: create_list(:question, 7, :textfield, section: @section2),
+      date: create_list(:question, 7, :date, section: @section3),
+      rda_metadata: create_list(:question, 7, :rda_metadata, section: @section1, options: 3),
+      checkbox: create_list(:question, 7, :checkbox, section: @section2, options: 3),
+      radiobutton: create_list(:question, 7, :radiobuttons, section: @section3, options: 3),
+      dropdown: create_list(:question, 7, :dropdown, section: @section1, options: 3),
+      multiselectbox: create_list(:question, 7, :multiselectbox, section: @section2, options: 3)
+    }
+  end
+
+  def create_answers
+    question_types_with_question_options = %i[checkbox radiobutton dropdown multiselectbox]
+    answers = {}
+    @non_conditional_questions.each do |question_type, questions|
+      answers[question_type] = questions.map do |question|
+        if question_types_with_question_options.include?(question_type)
+          create(:answer, plan: @plan, question: question, question_options: [question.question_options[2]], user: @user)
+        else
+          create(:answer, plan: @plan, question: question, user: @user)
+        end
+      end
+    end
+    answers
   end
 end