--- a/spec/controllers/answers_controller_with_conditional_questions_spec.rb
+++ b/spec/controllers/answers_controller_with_conditional_questions_spec.rb
@@ -289,17 +289,7 @@ RSpec.describe AnswersController, type: :controller do
         expect(ActionMailer::Base.deliveries.count).to eq(1)
         webhook_data = JSON.parse(add_webhook_condition.webhook_data)
 
-        ActionMailer::Base.deliveries.first do |mail|
-          expect(mail.to).to eq([webhook_data['email']])
-          expect(mail.subject).to eq(webhook_data['subject'])
-          expect(mail.body.encoded).to include(webhook_data['message'])
-          # To see structure of email sent see app/views/user_mailer/question_answered.html.erb.
-
-          # Message should have @user.name, chosen option text and question text.
-          expect(mail.body.encoded).to include(@user.name)
-          expect(mail.body.encoded).to include(@conditional_questions[:checkbox].question_options[2].text)
-          expect(mail.body.encoded).to include(@conditional_questions[:checkbox].text)
-        end
+        check_delivered_mail_for_webhook_data_and_question_data(webhook_data, :checkbox)
       end
       it 'handles multiple checkbox options (one of which is add_webhook condition)' do
         add_webhook_condition = create(:condition,
@@ -342,17 +332,7 @@ RSpec.describe AnswersController, type: :controller do
         expect(ActionMailer::Base.deliveries.count).to eq(1)
         webhook_data = JSON.parse(add_webhook_condition.webhook_data)
 
-        ActionMailer::Base.deliveries.first do |mail|
-          expect(mail.to).to eq([webhook_data['email']])
-          expect(mail.subject).to eq(webhook_data['subject'])
-          expect(mail.body.encoded).to include(webhook_data['message'])
-          # To see structure of email sent see app/views/user_mailer/question_answered.html.erb.
-
-          # Message should have @user.name, chosen option text and question text.
-          expect(mail.body.encoded).to include(@user.name)
-          expect(mail.body.encoded).to include(@conditional_questions[:checkbox].question_options[2].text)
-          expect(mail.body.encoded).to include(@conditional_questions[:checkbox].text)
-        end
+        check_delivered_mail_for_webhook_data_and_question_data(webhook_data, :checkbox)
       end
 
       it 'handles selection of a dropdown option (with add_webhook condition)' do
@@ -388,17 +368,7 @@ RSpec.describe AnswersController, type: :controller do
         expect(ActionMailer::Base.deliveries.count).to eq(1)
         webhook_data = JSON.parse(add_webhook_condition.webhook_data)
 
-        ActionMailer::Base.deliveries.first do |mail|
-          expect(mail.to).to eq([webhook_data['email']])
-          expect(mail.subject).to eq(webhook_data['subject'])
-          expect(mail.body.encoded).to include(webhook_data['message'])
-          # To see structure of email sent see app/views/user_mailer/question_answered.html.erb.
-
-          # Message should have @user.name, chosen option text and question text.
-          expect(mail.body.encoded).to include(@user.name)
-          expect(mail.body.encoded).to include(@conditional_questions[:dropdown].question_options[2].text)
-          expect(mail.body.encoded).to include(@conditional_questions[:dropdown].text)
-        end
+        check_delivered_mail_for_webhook_data_and_question_data(webhook_data, :dropdown)
       end
 
       it 'handles selection of a radiobutton option (with add_webhook condition)' do
@@ -434,17 +404,7 @@ RSpec.describe AnswersController, type: :controller do
         expect(ActionMailer::Base.deliveries.count).to eq(1)
         webhook_data = JSON.parse(add_webhook_condition.webhook_data)
 
-        ActionMailer::Base.deliveries.first do |mail|
-          expect(mail.to).to eq([webhook_data['email']])
-          expect(mail.subject).to eq(webhook_data['subject'])
-          expect(mail.body.encoded).to include(webhook_data['message'])
-          # To see structure of email sent see app/views/user_mailer/question_answered.html.erb.
-
-          # Message should have @user.name, chosen option text and question text.
-          expect(mail.body.encoded).to include(@user.name)
-          expect(mail.body.encoded).to include(@conditional_questions[:radiobutton].question_options[2].text)
-          expect(mail.body.encoded).to include(@conditional_questions[:radiobutton].text)
-        end
+        check_delivered_mail_for_webhook_data_and_question_data(webhook_data, :radiobutton)
       end
     end
   end
@@ -489,5 +449,19 @@ RSpec.describe AnswersController, type: :controller do
       end
     end
     answers
+  end
+
+  def check_delivered_mail_for_webhook_data_and_question_data(webhook_data, question_type)
+    ActionMailer::Base.deliveries.first do |mail|
+      expect(mail.to).to eq([webhook_data['email']])
+      expect(mail.subject).to eq(webhook_data['subject'])
+      expect(mail.body.encoded).to include(webhook_data['message'])
+      # To see structure of email sent see app/views/user_mailer/question_answered.html.erb.
+
+      # Message should have @user.name, chosen option text and question text.
+      expect(mail.body.encoded).to include(@user.name)
+      expect(mail.body.encoded).to include(@conditional_questions[question_type].question_options[2].text)
+      expect(mail.body.encoded).to include(@conditional_questions[question_type].text)
+    end
   end
 end