… spec/features/questions/conditions_questions_spec.rb
… spec/support/helpers/conditional_questions_helper.rb