Changes proposed in this PR:
- https://github.com/DMPRoadmap/roadmap/pull/3516 adds a lot of test coverage within `spec/features/questions/conditions_questions_spec.rb` and `spec/controllers/answers_controller_with_conditional_questions_spec.rb`. This PR attempts to refactor those files by applying DRY principles.
- `spec/support/helpers/conditional_questions_helper.rb` has been created to facilitate the shared logic between the two aforementioned files.
- The only actual change is the addition of `@non_conditional_questions[:radiobutton][index].id` to some `remove_data: ` values. (I am not sure why these `:radiobutton` options were absent in the first place and could add them back if needed.)