@aaronskiba 2025-06-02
The tests are getting noticeably faster, but I will try to improve things a bit more.

### BEFORE
https://github.com/DMPRoadmap/roadmap/actions/runs/14705500339/job/41264851295
```
Top 10 slowest example groups:
  AnswersController
    3.95 seconds average (43.5 seconds / 11 examples) ./spec/controllers/answers_controller_with_conditional_questions_spec.rb:5

Top 10 slowest example groups:
  Question::Conditions questions
    10.24 seconds average (92.2 seconds / 9 examples) ./spec/features/questions/conditions_questions_spec.rb:5
```

### AFTER
https://github.com/DMPRoadmap/roadmap/actions/runs/15404832747/job/43345375835?pr=3522 (up to commit [8893f32](https://github.com/DMPRoadmap/roadmap/pull/3522/commits/8893f32b3917dc4ef9f0f0bec077a6410311a374))
```
Top 10 slowest example groups:
  AnswersController
    1.98 seconds average (21.78 seconds / 11 examples) ./spec/controllers/answers_controller_with_conditional_questions_spec.rb:5

Top 10 slowest example groups:
  Templates::UpgradeCustomisations
    9.57 seconds average (9.57 seconds / 1 example) ./spec/features/templates/templates_upgrade_customisations_spec.rb:5
  Question::Conditions questions
    9.49 seconds average (85.44 seconds / 9 examples) ./spec/features/questions/conditions_questions_spec.rb:5
```