@aaronskiba 2025-06-04
Commit [e951b4d](https://github.com/DMPRoadmap/roadmap/pull/3522/commits/e951b4dd710eaa5dc1da1540f4a0f84ed467343c) greatly reduces the test time overhead. `./spec/features/questions/conditions_questions_spec.rb is now being executed 3x faster:

```
Top 10 slowest example groups:
  Question::Conditions questions
    9.36 seconds average (28.09 seconds / 3 examples) ./spec/features/questions/conditions_questions_spec.rb:5
```

However, the testing approach is a bit different with this change. Here is a copy of the commit description which describes the change:

- **Previously, `spec/features/questions/conditions_questions_spec.rb` tested `remove_data` and `add_webhook` functionality separately for each question type: checkbox, radiobutton, and dropdown. While thorough, this approach added a lot of overhead to test execution time.**
  **This change reduces that overhead by randomly choosing only one of checkbox, radiobutton, or dropdown for each of the tests. Over time, repeated executions via the GitHub Actions should ensure that all types continue to be exercised, maintaining coverage while also improving overall test time.**